refactor: split jules.js into smaller modules
Extracts functionality from the monolithic `jules.js` into four distinct, single-responsibility modules:
- `jules-keys.js`: Manages API key encryption and storage.
- `jules-queue.js`: Handles all queue-related data operations.
- `jules-modal.js`: Manages all modal UI rendering and event handlers.
- `jules-api.js`: Contains core API interaction logic with the Jules service.

This refactoring improves code organization and maintainability.

The change also resolves a circular dependency that was introduced between `jules-modal.js` and `jules-queue.js` by ensuring that UI event handler attachment is solely the responsibility of the modal module. All necessary imports in `app.js` and `prompt-renderer.js` have been updated to reflect the new file structure.
