+++ src/modules/jules-keys.js
@@ -1,0 +1,44 @@
+export async function checkJulesKey(uid) {
+  try {
+    if (!window.db) {
+      return false;
+    }
+    const doc = await window.db.collection('julesKeys').doc(uid).get();
+    return doc.exists;
+  } catch (error) {
+    return false;
+  }
+}
+
+export async function deleteStoredJulesKey(uid) {
+  try {
+    if (!window.db) return false;
+    await window.db.collection('julesKeys').doc(uid).delete();
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
+export async function encryptAndStoreKey(plaintext, uid) {
+  try {
+    const paddedUid = (uid + '\0'.repeat(32)).slice(0, 32);
+    const keyData = new TextEncoder().encode(paddedUid);
+    const key = await window.crypto.subtle.importKey('raw', keyData, { name: 'AES-GCM' }, false, ['encrypt']);
+
+    const ivString = uid.slice(0, 12).padEnd(12, '0');
+    const iv = new TextEncoder().encode(ivString).slice(0, 12);
+    const plaintextData = new TextEncoder().encode(plaintext);
+    const ciphertext = await window.crypto.subtle.encrypt({ name: 'AES-GCM', iv }, key, plaintextData);
+    const encrypted = btoa(String.fromCharCode(...new Uint8Array(ciphertext)));
+
+    if (!window.db) throw new Error('Firestore not initialized');
+    await window.db.collection('julesKeys').doc(uid).set({
+      key: encrypted,
+      storedAt: firebase.firestore.FieldValue.serverTimestamp()
+    });
+    return true;
+  } catch (error) {
+    throw error;
+  }
+}
